requirements.txt
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.0
pymde==0.1.18
scikit-learn==1.1.2
seaborn==0.12.0
umap-learn==0.5.3
tensorflow==2.11.0
tensorflow_hub==0.12.0
wordcloud==1.8.2.2